Fix Wsign-compare

+++ src/LLGI.Graphics.h
@@ -110,7 +110,7 @@ struct RenderPassPipelineStateKey
 			ret += std::hash<bool>()(key.HasResolvedRenderTarget);
 			ret += std::hash<bool>()(key.HasResolvedDepthTarget);
 
-			for (int32_t i = 0; i < key.RenderTargetFormats.size(); i++)
+			for (size_t i = 0; i < key.RenderTargetFormats.size(); i++)
 			{
 				ret += std::hash<uint64_t>()((uint64_t)key.RenderTargetFormats.at(i));
 			}
